feat(vim.ui.open): allow configurable opener in vim.ui.open
With the welcomed demise of `netrw`, we lost an opportunity
to use an arbitrary command for opening URL-like strings
with `gx`. This should suggest a way how to recover this
functionality.

Of course, new name of the variable `vim.g.netrw_browsex_viewer`
should be developed (I am not sure what are the standards for
creating new options).

Another thing which is missing are tests. I tried to create one,
but I got completely lost and frankly `neovim` is not my primary
editor, so I don’t have a time to dive too deep into its innards.

Will eventually Fix: https://github.com/neovim/neovim/issues/29488